Fluff C3 seam 1: extract text/serialisation/local-draft utilities to editorText.ts
Verbatim move of the whitespace-normalised search, plain-text
serialisation, outline/stats, and localStorage draft helpers out of
DocumentRichEditor.tsx. DocumentRichEditor re-exports every previously
public name, so DocumentDetail and the tests keep their import path.
